install lib moment

+++ src/routes/car.route.ts
@@ -1,6 +1,5 @@
 import { Request, Response, Router } from "express";
 import CarService from "../services/car.service";
-const ObjectId = require("mongoose").Types.ObjectId;
 
 const router = Router();
 
@@ -8,11 +7,8 @@ router.get("/", async (req: Request, res: Response) => {
   const cars = await CarService.getAll();
   res.status(200).send(cars);
 });
-router.get("/:document", async (req: Request, res: Response) => {
-  console.log(req.params.id);
-
-  const id = new ObjectId(req.params.id);
-  const car = await CarService.getById(id);
+router.get("/:id", async (req: Request, res: Response) => {
+  const car = await CarService.getById(req.params.id);
   if (!car) {
     return res.status(404).send({ message: "Carro não encontrado!" });
   }
@@ -22,31 +18,40 @@ router.post("/", async (req: Request, res: Response) => {
   try {
     await CarService.create(req.body);
     res.status(201).send({ message: "Carro criado com sucesso!" });
+  } catch (error: any) {
+    res.status(400).send({
+      message: error.message,
+    });
+  }
+});
+router.put("/:id", async (req: Request, res: Response) => {
+  try {
+    const car = await CarService.getById(req.params.id);
+    if (!car) {
+      return res.status(404).send({ message: "Carro não encontrado!" });
+    }
+    await CarService.update(req.params.id, req.body);
+    res.status(200).send({ message: "Carro atualizado com sucesso!" });
+  } catch {
+    res.status(400).send({
+      messsage: "Aconteceu algo, favor verificar as informações enviadas.",
+    });
+  }
+});
+router.delete("/:id", async (req: Request, res: Response) => {
+  try {
+    const car = await CarService.getById(req.params.id);
+    if (!car) {
+      return res.status(404).send({ message: "Carro não encontrado!" });
+    }
+    const at = await CarService.delete(req.params.id);
+    console.log(at);
+    res.status(200).send({ message: "Carro excluído com sucesso!" });
   } catch {
     res.status(400).send({
       message: "Aconteceu algo, favor verificar as informações enviadas.",
     });
   }
-  router.put("/:document", async (req: Request, res: Response) => {
-    try {
-      await CarService.update(req.params.id, req.body);
-      res.status(200).send({ message: "Carro atualizado com sucesso!" });
-    } catch {
-      res.status(400).send({
-        messsage: "Aconteceu algo, favor verificar as informações enviadas.",
-      });
-    }
-  });
-  router.delete("/:document", async (req: Request, res: Response) => {
-    try {
-      await CarService.delete(req.params.id);
-      res.status(200).send({ message: "Carro excluído com sucesso!" });
-    } catch {
-      res.status(400).send({
-        message: "Aconteceu algo, favor verificar as informações enviadas.",
-      });
-    }
-  });
 });
 
 export default router;
