+++ src/routes/user.route.ts
@@ -7,7 +7,7 @@ router.get("/", async (req: Request, res: Response) => {
   const users = await UserService.getAll();
   res.status(200).send(users);
 });
-router.get("/:document", async (req: Request, res: Response) => {
+router.get("/:id", async (req: Request, res: Response) => {
   const user = await UserService.getById(req.params.id);
   console.log(user);
   if (!user) {
@@ -19,31 +19,50 @@ router.post("/", async (req: Request, res: Response) => {
   try {
     await UserService.createUser(req.body);
     res.status(201).send({ message: "Usuário criado com sucesso!" });
+  } catch (error: any) {
+    res.status(400).send({
+      message: error.message,
+    });
+  }
+});
+router.post("/auth", async (req: Request, res: Response) => {
+  const { email, password } = req.body;
+
+  try {
+    const token = await UserService.auth(email, password);
+    res.status(200).send({ token });
+  } catch (error: any) {
+    res.status(401).send({ message: error.message });
+  }
+});
+
+router.put("/:id", async (req: Request, res: Response) => {
+  try {
+    const user = await UserService.getById(req.params.id);
+    if (!user) {
+      return res.status(404).send({ message: "Usuário não encontrado!" });
+    }
+    await UserService.updateUser(req.params.id, req.body);
+    res.status(200).send({ message: "Usuário atualizado com sucesso!" });
+  } catch {
+    res.status(400).send({
+      messsage: "Aconteceu algo, favor verificar as informações enviadas.",
+    });
+  }
+});
+router.delete("/:id", async (req: Request, res: Response) => {
+  try {
+    const user = await UserService.getById(req.params.id);
+    if (!user) {
+      return res.status(404).send({ message: "Usuário não encontrado!" });
+    }
+    await UserService.deleteUser(req.params.id);
+    res.status(200).send({ message: "Usuário excluído com sucesso!" });
   } catch {
     res.status(400).send({
       message: "Aconteceu algo, favor verificar as informações enviadas.",
     });
   }
-  router.put("/:document", async (req: Request, res: Response) => {
-    try {
-      await UserService.updateUser(req.params.id, req.body);
-      res.status(200).send({ message: "Usuário atualizado com sucesso!" });
-    } catch {
-      res.status(400).send({
-        messsage: "Aconteceu algo, favor verificar as informações enviadas.",
-      });
-    }
-  });
-  router.delete("/:document", async (req: Request, res: Response) => {
-    try {
-      await UserService.deleteUser(req.params.id);
-      res.status(200).send({ message: "Usuário excluído com sucesso!" });
-    } catch {
-      res.status(400).send({
-        message: "Aconteceu algo, favor verificar as informações enviadas.",
-      });
-    }
-  });
 });
 
 export default router;
